push

+++ payments/models.py
@@ -2,7 +2,7 @@ from accounts.models import TelegramUser
 from django.db import models
 class Wallet(models.Model):
     user = models.OneToOneField(TelegramUser, on_delete=models.CASCADE, related_name='wallet')
-    balance = models.DecimalField(max_digits=10, decimal_places=2, default=0.00)
+    balance = models.FloatField(default=0.00)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
